requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

ultralytics==8.1.0
torch==2.2.0
torchvision==0.17.0
opencv-python-headless==4.9.0.80
pillow==10.1.0
numpy==1.24.3

python-dotenv==1.0.0